+++ src/shm_ext.c
@@ -140,7 +140,7 @@ sr_shmext_conn_remap_unlock(sr_conn_ctx_t *conn, sr_lock_mode_t mode, int ext_lo
             last = iter;
         }
 
-        if (last && (((char *)last - conn->ext_shm.addr) + last->size == (signed)conn->ext_shm.size)) {
+        if (last && ((uint32_t)((char *)last - conn->ext_shm.addr) + last->size == conn->ext_shm.size)) {
             if ((err_info = sr_file_get_size(conn->ext_shm.fd, &shm_file_size))) {
                 goto cleanup_unlock;
             }
@@ -479,7 +479,7 @@ sr_shmext_print(sr_mod_shm_t *mod_shm, sr_shm_t *shm_ext)
 
         /* check alignment */
         assert(items[i].size == SR_SHM_SIZE(items[i].size));
-        assert((unsigned)items[i].start == SR_SHM_SIZE(items[i].start));
+        assert(items[i].start == SR_SHM_SIZE(items[i].start));
 
         /* free name */
         free(items[i].name);
